Conflicts:
	gtk/Makefile.am

+++ gdk/quartz/gdkcursor-quartz.c
@@ -84,7 +84,7 @@ create_builtin_cursor (GdkCursorType cursor_type)
 {
   GdkCursor *cursor;
   NSBitmapImageRep *bitmap_rep;
-  gint mask_width, mask_height;
+  NSInteger mask_width, mask_height;
   gint src_width, src_height;
   gint dst_stride;
   const guchar *mask_start, *src_start;
@@ -249,7 +249,7 @@ gdk_cursor_new_from_pixmap (GdkPixmap      *source,
   NSImage *image;
   NSCursor *nscursor;
   GdkCursor *cursor;
-  gint width, height;
+  int width, height;
   gint tmp_x, tmp_y;
   guchar *dst_data, *mask_data, *src_data;
   guchar *mask_start, *src_start;
@@ -265,7 +265,7 @@ gdk_cursor_new_from_pixmap (GdkPixmap      *source,
   gdk_drawable_get_size (source, &width, &height);
 
   bitmap_rep = [[NSBitmapImageRep alloc] initWithBitmapDataPlanes:NULL
-		pixelsWide:width pixelsHigh:height
+		pixelsWide:(NSInteger)width pixelsHigh:(NSInteger)height
 		bitsPerSample:8 samplesPerPixel:4
 		hasAlpha:YES isPlanar:NO colorSpaceName:NSDeviceRGBColorSpace
 		bytesPerRow:0 bitsPerPixel:0];
